Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main(){
-    float n;
-    scanf("%f", &n);
+    int n;
+    scanf("%d", &n);
     printf("You entered: ",n );
     return 0;
